backend/requirements.txt
paho-mqtt
psycopg2-binary
fastapi
uvicorn
pandas
streamlit
requests